+++ prov/verbs/Makefile.include
@@ -9,6 +9,7 @@ _verbs_files =							\
 	prov/verbs/src/verbs_eq.c				\
 	prov/verbs/src/verbs_info.c				\
 	prov/verbs/src/verbs_ep.c				\
+	prov/verbs/src/verbs_msg.c				\
 	prov/verbs/src/ep_rdm/verbs_av_ep_rdm.c			\
 	prov/verbs/src/ep_rdm/verbs_cq_ep_rdm.c			\
 	prov/verbs/src/ep_rdm/verbs_ep_rdm.c			\
