+++ prov/verbs/src/verbs_ep.c
@@ -811,179 +811,6 @@ err1:
 	return ret;
 }
 
-static inline ssize_t
-fi_ibv_msg_ep_recvmsg(struct fid_ep *ep_fid, const struct fi_msg *msg, uint64_t flags)
-{
-	struct fi_ibv_ep *ep =
-		container_of(ep_fid, struct fi_ibv_ep, util_ep.ep_fid);
-	struct ibv_recv_wr wr = {
-		.wr_id = (uintptr_t)msg->context,
-		.num_sge = msg->iov_count,
-		.next = NULL,
-	};
-	struct ibv_recv_wr *bad_wr;
-
-	assert(ep->util_ep.rx_cq);
-
-	fi_ibv_set_sge_iov(wr.sg_list, msg->msg_iov, msg->iov_count, msg->desc);
-
-	return fi_ibv_handle_post(ibv_post_recv(ep->id->qp, &wr, &bad_wr));
-}
-
-static ssize_t
-fi_ibv_msg_ep_recv(struct fid_ep *ep_fid, void *buf, size_t len,
-		void *desc, fi_addr_t src_addr, void *context)
-{
-	struct fi_ibv_ep *ep =
-		container_of(ep_fid, struct fi_ibv_ep, util_ep.ep_fid);
-	struct ibv_sge sge = fi_ibv_init_sge(buf, len, desc);
-	struct ibv_recv_wr wr = {
-		.wr_id = (uintptr_t)context,
-		.num_sge = 1,
-		.sg_list = &sge,
-		.next = NULL,
-	};
-	struct ibv_recv_wr *bad_wr;
-
-	assert(ep->util_ep.rx_cq);
-
-	return fi_ibv_handle_post(ibv_post_recv(ep->id->qp, &wr, &bad_wr));
-}
-
-static ssize_t
-fi_ibv_msg_ep_recvv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-                 size_t count, fi_addr_t src_addr, void *context)
-{
-	struct fi_msg msg = {
-		.msg_iov = iov,
-		.desc = desc,
-		.iov_count = count,
-		.addr = src_addr,
-		.context = context,
-	};
-
-	return fi_ibv_msg_ep_recvmsg(ep_fid, &msg, 0);
-}
-
-static ssize_t
-fi_ibv_msg_ep_sendmsg(struct fid_ep *ep_fid, const struct fi_msg *msg, uint64_t flags)
-{
-	struct fi_ibv_ep *ep =
-		container_of(ep_fid, struct fi_ibv_ep, util_ep.ep_fid);
-	struct ibv_send_wr wr = {
-		.wr_id = (uintptr_t)msg->context,
-	};
-
-	if (flags & FI_REMOTE_CQ_DATA) {
-		wr.opcode = IBV_WR_SEND_WITH_IMM;
-		wr.imm_data = htonl((uint32_t)msg->data);
-	} else {
-		wr.opcode = IBV_WR_SEND;
-	}
-
-	return fi_ibv_send_msg(ep, &wr, msg, flags);
-}
-
-static ssize_t
-fi_ibv_msg_ep_send(struct fid_ep *ep_fid, const void *buf, size_t len,
-		void *desc, fi_addr_t dest_addr, void *context)
-{
-	struct fi_ibv_ep *ep =
-		container_of(ep_fid, struct fi_ibv_ep, util_ep.ep_fid);
-	struct ibv_send_wr wr = {
-		.wr_id = (uintptr_t)context,
-		.opcode = IBV_WR_SEND,
-		.send_flags = VERBS_INJECT(ep, len) | VERBS_COMP(ep),
-	};
-
-	return fi_ibv_send_buf(ep, &wr, buf, len, desc);
-}
-
-static ssize_t
-fi_ibv_msg_ep_senddata(struct fid_ep *ep_fid, const void *buf, size_t len,
-		       void *desc, uint64_t data, fi_addr_t dest_addr, void *context)
-{
-	struct fi_ibv_ep *ep =
-		container_of(ep_fid, struct fi_ibv_ep, util_ep.ep_fid);
-	struct ibv_send_wr wr = {
-		.wr_id = (uintptr_t)context,
-		.opcode = IBV_WR_SEND_WITH_IMM,
-		.imm_data = htonl((uint32_t)data),
-		.send_flags = VERBS_INJECT(ep, len) | VERBS_COMP(ep),
-	};
-
-	return fi_ibv_send_buf(ep, &wr, buf, len, desc);
-}
-
-static ssize_t
-fi_ibv_msg_ep_sendv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-		    size_t count, fi_addr_t dest_addr, void *context)
-{
-	struct fi_ibv_ep *ep =
-		container_of(ep_fid, struct fi_ibv_ep, util_ep.ep_fid);
-	struct ibv_send_wr wr = {
-		.wr_id = (uintptr_t)context,
-		.opcode = IBV_WR_SEND,
-	};
-
-	return fi_ibv_send_iov(ep, &wr, iov, desc, count);
-}
-
-static ssize_t fi_ibv_msg_ep_inject(struct fid_ep *ep_fid, const void *buf, size_t len,
-		fi_addr_t dest_addr)
-{
-	struct fi_ibv_ep *ep =
-		container_of(ep_fid, struct fi_ibv_ep, util_ep.ep_fid);
-	struct ibv_send_wr wr = {
-		.wr_id = VERBS_INJECT_FLAG,
-		.opcode = IBV_WR_SEND,
-		.send_flags = IBV_SEND_INLINE,
-	};
-
-	return fi_ibv_send_buf_inline(ep, &wr, buf, len);
-}
-
-static ssize_t fi_ibv_msg_ep_injectdata(struct fid_ep *ep_fid, const void *buf, size_t len,
-		    uint64_t data, fi_addr_t dest_addr)
-{
-	struct fi_ibv_ep *ep =
-		container_of(ep_fid, struct fi_ibv_ep, util_ep.ep_fid);
-	struct ibv_send_wr wr = {
-		.wr_id = VERBS_INJECT_FLAG,
-		.opcode = IBV_WR_SEND_WITH_IMM,
-		.imm_data = htonl((uint32_t)data),
-		.send_flags = IBV_SEND_INLINE,
-	};
-
-	return fi_ibv_send_buf_inline(ep, &wr, buf, len);
-}
-
-struct fi_ops_msg fi_ibv_msg_ep_msg_ops = {
-	.size = sizeof(struct fi_ops_msg),
-	.recv = fi_ibv_msg_ep_recv,
-	.recvv = fi_ibv_msg_ep_recvv,
-	.recvmsg = fi_ibv_msg_ep_recvmsg,
-	.send = fi_ibv_msg_ep_send,
-	.sendv = fi_ibv_msg_ep_sendv,
-	.sendmsg = fi_ibv_msg_ep_sendmsg,
-	.inject = fi_ibv_msg_ep_inject,
-	.senddata = fi_ibv_msg_ep_senddata,
-	.injectdata = fi_ibv_msg_ep_injectdata,
-};
-
-struct fi_ops_msg fi_ibv_msg_srq_ep_msg_ops = {
-	.size = sizeof(struct fi_ops_msg),
-	.recv = fi_no_msg_recv,
-	.recvv = fi_no_msg_recvv,
-	.recvmsg = fi_no_msg_recvmsg,
-	.send = fi_ibv_msg_ep_send,
-	.sendv = fi_ibv_msg_ep_sendv,
-	.sendmsg = fi_ibv_msg_ep_sendmsg,
-	.inject = fi_ibv_msg_ep_inject,
-	.senddata = fi_ibv_msg_ep_senddata,
-	.injectdata = fi_ibv_msg_ep_injectdata,
-};
-
 static struct fi_ops_ep fi_ibv_srq_ep_base_ops = {
 	.size = sizeof(struct fi_ops_ep),
 	.cancel = fi_no_cancel,
